Refactor workflows management routes and consolidate route definitions
