+++ src/platform/plugins/shared/workflows_management/server/workflows_management/routes/get_workflow_aggs.ts
@@ -1,0 +1,48 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { schema } from '@kbn/config-schema';
+import type { RouteDependencies } from './types';
+
+export function registerGetWorkflowAggsRoute({ router, api, logger, spaces }: RouteDependencies) {
+  router.get(
+    {
+      path: '/api/workflows/aggs',
+      options: {
+        tags: ['api', 'workflows'],
+      },
+      security: {
+        authz: {
+          requiredPrivileges: [
+            {
+              anyRequired: ['read', 'workflow_read'],
+            },
+          ],
+        },
+      },
+      validate: { query: schema.object({ fields: schema.arrayOf(schema.string()) }) },
+    },
+    async (context, request, response) => {
+      try {
+        const { fields } = request.query as { fields: string[] };
+        const spaceId = spaces.getSpaceId(request);
+        const aggs = await api.getWorkflowAggs(fields, spaceId);
+
+        return response.ok({ body: aggs || {} });
+      } catch (error) {
+        return response.customError({
+          statusCode: 500,
+          body: {
+            message: `Internal server error: ${error}`,
+          },
+        });
+      }
+    }
+  );
+}
