+++ src/platform/plugins/shared/workflows_management/server/workflows_management/routes/get_workflow_executions.ts
@@ -1,0 +1,109 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { type Type, schema } from '@kbn/config-schema';
+import type { ExecutionStatus, ExecutionType } from '@kbn/workflows';
+import { ExecutionStatusValues, ExecutionTypeValues } from '@kbn/workflows';
+import type { SearchWorkflowExecutionsParams } from '../workflows_management_service';
+import { parseExecutionStatuses, MAX_PAGE_SIZE } from './types';
+import type { RouteDependencies } from './types';
+
+export function registerGetWorkflowExecutionsRoute({
+  router,
+  api,
+  logger,
+  spaces,
+}: RouteDependencies) {
+  router.get(
+    {
+      path: '/api/workflowExecutions',
+      options: {
+        tags: ['api', 'workflows'],
+      },
+      security: {
+        authz: {
+          requiredPrivileges: [
+            {
+              anyRequired: ['read', 'workflow_execution_read'],
+            },
+          ],
+        },
+      },
+      validate: {
+        // todo use shared params schema based on SearchWorkflowExecutionsParams type
+        query: schema.object({
+          workflowId: schema.string(),
+          statuses: schema.maybe(
+            schema.oneOf(
+              [
+                schema.oneOf(
+                  ExecutionStatusValues.map((type) => schema.literal(type)) as [
+                    Type<ExecutionStatus>
+                  ]
+                ),
+                schema.arrayOf(
+                  schema.oneOf(
+                    ExecutionStatusValues.map((type) => schema.literal(type)) as [
+                      Type<ExecutionStatus>
+                    ]
+                  )
+                ),
+              ],
+              {
+                defaultValue: [],
+              }
+            )
+          ),
+          executionTypes: schema.maybe(
+            schema.oneOf(
+              [
+                schema.oneOf(
+                  ExecutionTypeValues.map((type) => schema.literal(type)) as [Type<ExecutionType>]
+                ),
+                schema.arrayOf(
+                  schema.oneOf(
+                    ExecutionTypeValues.map((type) => schema.literal(type)) as [Type<ExecutionType>]
+                  )
+                ),
+              ],
+              {
+                defaultValue: [],
+              }
+            )
+          ),
+          page: schema.maybe(schema.number({ min: 1 })),
+          perPage: schema.maybe(schema.number({ min: 1, max: MAX_PAGE_SIZE })),
+        }),
+      },
+    },
+    async (context, request, response) => {
+      try {
+        const spaceId = spaces.getSpaceId(request);
+        const params: SearchWorkflowExecutionsParams = {
+          workflowId: request.query.workflowId,
+          statuses: parseExecutionStatuses(request.query.statuses),
+          // Execution type filter is not supported yet
+          // executionTypes: parseExecutionTypes(request.query.executionTypes),
+          page: request.query.page,
+          perPage: request.query.perPage,
+        };
+        return response.ok({
+          body: await api.getWorkflowExecutions(params, spaceId),
+        });
+      } catch (error) {
+        return response.customError({
+          statusCode: 500,
+          body: {
+            message: `Internal server error: ${error}`,
+          },
+        });
+      }
+    }
+  );
+}
