+++ src/platform/plugins/shared/workflows_management/server/workflows_management/routes/put_update_workflow.ts
@@ -1,0 +1,64 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { schema } from '@kbn/config-schema';
+import { UpdateWorkflowCommandSchema } from '@kbn/workflows';
+import { isWorkflowValidationError } from '../../../common/lib/errors';
+import type { RouteDependencies } from './types';
+
+export function registerPutUpdateWorkflowRoute({ router, api, logger, spaces }: RouteDependencies) {
+  router.put(
+    {
+      path: '/api/workflows/{id}',
+      options: {
+        tags: ['api', 'workflows'],
+      },
+      security: {
+        authz: {
+          requiredPrivileges: [
+            {
+              anyRequired: ['all', 'workflow_update'],
+            },
+          ],
+        },
+      },
+      validate: {
+        params: schema.object({
+          id: schema.string(),
+        }),
+        body: UpdateWorkflowCommandSchema.partial(),
+      },
+    },
+    async (context, request, response) => {
+      try {
+        const { id } = request.params as { id: string };
+        const spaceId = spaces.getSpaceId(request);
+        const updated = await api.updateWorkflow(id, request.body, spaceId, request);
+        if (updated === null) {
+          return response.notFound();
+        }
+        return response.ok({
+          body: updated,
+        });
+      } catch (error) {
+        if (isWorkflowValidationError(error)) {
+          return response.badRequest({
+            body: error.toJSON(),
+          });
+        }
+        return response.customError({
+          statusCode: 500,
+          body: {
+            message: `Internal server error: ${error}`,
+          },
+        });
+      }
+    }
+  );
+}
